+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Unreleased
 ----------
 
 - Fixed a bug where visiting ``/<database>/-/query`` without a ``?sql=`` parameter returned a 500 error. (:issue:`2743`)
+- The ``top_canned_query()`` plugin hook has been renamed to :ref:`top_stored_query() <plugin_hook_top_stored_query>`. (:issue:`2747`)
 
 .. _v1_0_a30: